Django_practice/DjangoAdvance03/requirements.txt
django-redis
django-redis-cache
djangorestframework
django-filter